refactor: CoordinateKeys 패키지 이동
