+++ android/app/src/main/java/io/coursepick/coursepick/presentation/CoordinateKeys.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.util
+package io.coursepick.coursepick.presentation
 
 object CoordinateKeys {
     const val EXTRA_KEYS_LATITUDE = "latitude"
