+++ android/app/src/main/java/io/coursepick/coursepick/presentation/course/CoursesActivity.kt
@@ -29,6 +29,7 @@ import io.coursepick.coursepick.databinding.ActivityCoursesBinding
 import io.coursepick.coursepick.domain.Coordinate
 import io.coursepick.coursepick.domain.Latitude
 import io.coursepick.coursepick.domain.Longitude
+import io.coursepick.coursepick.presentation.CoordinateKeys
 import io.coursepick.coursepick.presentation.DoublePressDetector
 import io.coursepick.coursepick.presentation.map.kakao.KakaoMapManager
 import io.coursepick.coursepick.presentation.map.kakao.toCoordinate
@@ -37,7 +38,6 @@ import io.coursepick.coursepick.presentation.preference.PreferencesActivity
 import io.coursepick.coursepick.presentation.routetfinder.RouteFinderApplication
 import io.coursepick.coursepick.presentation.routetfinder.RouteFinderChoiceDialogFragment
 import io.coursepick.coursepick.presentation.search.SearchActivity
-import io.coursepick.coursepick.util.CoordinateKeys
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
