+++ android/app/src/main/java/io/coursepick/coursepick/presentation/search/SearchActivity.kt
@@ -12,8 +12,8 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import io.coursepick.coursepick.databinding.ActivitySearchBinding
 import io.coursepick.coursepick.domain.Place
+import io.coursepick.coursepick.presentation.CoordinateKeys
 import io.coursepick.coursepick.presentation.Logger
-import io.coursepick.coursepick.util.CoordinateKeys
 
 class SearchActivity : AppCompatActivity() {
     private val binding by lazy { ActivitySearchBinding.inflate(layoutInflater) }
